sync dev

+++ packages/graphics/mesa/package.mk
@@ -10,17 +10,18 @@ PKG_TOOLCHAIN="meson"
 PKG_PATCH_DIRS+=" ${DEVICE}"
 
 case ${DEVICE} in
-#  RK3588*)
-#	PKG_VERSION="120202c675749c5ef81ae4c8cdc30019b4de08f4"
-#	PKG_SITE="https://gitlab.com/panfork/mesa"
-#	PKG_URL="${PKG_SITE}.git"
-#	PKG_GIT_CLONE_BRANCH="csf"
-#  ;;
-  RK3*) #Using upstream dev for panfrost
-	PKG_VERSION="db29984c254f60f5daeec0ea4e6048b6ee7902f8"
-	PKG_SITE="https://gitlab.freedesktop.org/mesa/mesa"
-	PKG_URL="${PKG_SITE}.git"
-	PKG_PATCH_DIRS+=" panfrost"
+  RK3*|S922X)
+    if [ "${DEVICE}" = "S922X" -a "${USE_MALI}" != "no" ]; then
+      PKG_VERSION="24.0.4"
+	    PKG_SITE="http://www.mesa3d.org/"
+	    PKG_URL="https://gitlab.freedesktop.org/mesa/mesa/-/archive/mesa-${PKG_VERSION}/mesa-mesa-${PKG_VERSION}.tar.gz"
+    else
+      #Using upstream dev for panfrost
+	    PKG_VERSION="7d4c23991a4cfa2cdc90315c736bf3f70a5f0238"
+	    PKG_SITE="https://gitlab.freedesktop.org/mesa/mesa"
+	    PKG_URL="${PKG_SITE}.git"
+	    PKG_PATCH_DIRS+=" panfrost"
+    fi
   ;;
   *)
 	PKG_VERSION="24.0.4"
@@ -106,9 +107,7 @@ else
 fi
 
 post_makeinstall_target() {
-  case ${DEVICE} in
-    S922X)
-      rm -f ${INSTALL}/usr/lib/libvulkan_panfrost.so ${INSTALL}/usr/share/vulkan/icd.d/panfrost_icd.aarch64.json
-    ;;
-  esac
+  if [ "${DEVICE}" = "S922X" -a "${USE_MALI}" != "no" ]; then
+    rm -f ${INSTALL}/usr/lib/libvulkan_panfrost.so ${INSTALL}/usr/share/vulkan/icd.d/panfrost_icd.aarch64.json
+  fi
 }
